# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Sistema de Controle Antimicrobianos Veterinários",
    "description": "Arquitetura de um Sistema de Controle de Uso, Vendas e Prescrição de Antimicrobianos Veterinários"
  },
  "security": [{"bearer":[]}],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/userPosition": {
      "get": {
        "summary": "Show all Oms classification",
        "tags": [
          "Oms Classification"
        ],
        "responses": {
          "200": {
            "description": "Return all Oms classification"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create an new Oms classification",
        "tags": [
          "Oms Classification"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/UserPosition/Create"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Oms classification created"
          },
          "400": {
            "description": "Error when creating oms classification"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Update an Oms classification",
        "tags": [
          "Oms Classification"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/UserPosition/Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Oms classification Deleted"
          },
          "400": {
            "description": "Error when deleting oms classification"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Error id dont was found"
          }
        }
      },
      "delete": {
        "summary": "Delete a oms classification",
        "tags": [
          "Oms Classification"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/UserPosition/Delete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Oms classification updated"
          },
          "400": {
            "description": "Error when delete oms classification"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}